add test for section_names
