+++ tests/unit-ngv/test_radii_change.py
@@ -5,7 +5,7 @@ import numpy.testing as npt
 
 from tests.conftest import NGV_DIR
 from neurodamus import Neurodamus
-from neurodamus.ngv import GlioVascularManager, GlutList
+from neurodamus.ngv import GlioVascularManager
 
 
 def get_Rad(astro_id, manager):
@@ -20,24 +20,6 @@ def get_R0pas(astro_id, manager):
     if astrocyte.endfeet is None:
         return []
     return [sec(0.5).vascouplingB.R0pas for sec in astrocyte.endfeet]
-
-
-def test_glut_list():
-    """Test base functionality of glut_list"""
-
-    ll = GlutList(range(5), -1)
-    assert ll == [0, 1, 2, 3, 4, -1]
-    ll.pop()
-    assert ll == [0, 1, 2, 3, -1]
-    ll.append(10)
-    assert ll == [0, 1, 2, 3, 10, -1]
-    ll[2] = 100
-    assert ll == [0, 1, 100, 3, 10, -1]
-    ll[-1] = -2
-    assert ll == [0, 1, 100, 3, 10, -2]
-    assert ll.tail == -2
-    ll.tail = -3
-    assert ll == [0, 1, 100, 3, 10, -3]
 
 
 @pytest.mark.parametrize("create_tmp_simulation_config_file", [
