+++ tox.ini
@@ -35,6 +35,23 @@ commands =
     {toxinidir}/ci/build_ndcore.sh {toxinidir}/neurodamus/data {toxworkdir}
     sh -c ". {toxworkdir}/.envfile && pytest -x --cov=neurodamus --cov-report=term-missing --forked {posargs:tests/unit}"
 
+[testenv:removeme]
+deps =
+    {[testenv]deps}
+    pytest-cov
+    neuron-nightly
+    morphio
+setenv =
+    NEURON_INIT_MPI=0
+    PYTHONPATH={toxinidir}
+    COVERAGE_FILE = .coverage.unit
+allowlist_externals =
+    {toxinidir}/ci/build_ndcore.sh
+    sh
+commands =
+    {toxinidir}/ci/build_ndcore.sh {toxinidir}/neurodamus/data {toxworkdir}
+    sh -c ". {toxworkdir}/.envfile && pytest -x --forked {posargs:tests/unit/test_morphIOwrapper.py}"
+
 [testenv:unit-ngv]
 deps =
     {[testenv]deps}
